+++ website/docs/components/select/examples/custom_selected_label.tsx
@@ -7,16 +7,22 @@ const options = Array(6)
   .fill('')
   .map((_, index) => ({
     value: index, // value of the selected option
-    label: `Option ${index}`, // the value displayed in the trigger when the option is selected
+    label: `Label ${index}`, // the value displayed in the trigger when the option is selected
     children: `Option ${index}`, // option's children displayed in the dropdown
   }));
 
 const Demo = () => (
   <Flex direction='column'>
-    <Text tag='label' size={200} htmlFor='basic-select'>
-      Basic select
+    <Text tag='label' size={200} htmlFor='select-custom-label'>
+      Select with custom selected label
     </Text>
-    <Select mt={2} mr='auto' options={options} placeholder='Select option' id='basic-select' />
+    <Select
+      mt={2}
+      mr='auto'
+      options={options}
+      placeholder='Select option'
+      id='select-custom-label'
+    />
   </Flex>
 );