+++ semcore/dropdown/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.38.0] - 2024-09-24
+
+### Added
+
+- `AbstractDropdown` internal class. It's needed to create DropdownMenu and Select.
+- `Dropdown.Item` with styles for DropdownMenu.Item and Select.Item.
+- `Dropdown.Group` with styles for wrap DropdownMenu.Item and Select.Item into groups.
+
 ## [4.37.1] - 2024-09-27
 
 ### Changed